+++ cf/mixin/fielddomainlist.py
@@ -331,3 +331,35 @@ class FieldDomainList:
         """
 
         return type(self)(f for f in self if f.match_by_rank(*ranks))
+
+    def to_xarray(self, group=True):
+        """Convert the list elements to an `xarray` Dataset.
+
+        {{cf_xarray description}}
+
+        Note that ``ds = fl.to_xarray()`` is identical to ``ds =
+        cf.write(fl, fmt='XARRAY')``. Also, `cfdm.write` allows a
+        mixture a mixture of fields and domains to be written to the
+        same `xarray` dataset.
+
+        .. versionadded:: NEXTVERSION
+
+        .. seealso:: `cf.write`
+
+        :Parameter:
+
+            group: `bool`, optional
+                If False then create a "flat" dataset, i.e. one with
+                only the root group, regardless of any group structure
+                specified by the netCDF interfaces of the list
+                elements and their components. If True (the default)
+                then any sub-groups will be created and populated.
+
+        :Returns:
+
+            {{Returns xarray}}
+
+        """
+        from cf.read_write import write
+
+        return write(self, fmt="XARRAY", group=group)