+++ Changelog.rst
@@ -3,6 +3,13 @@ Version 3.20.1
 
 **2026-06-29**
 
+* New methods to convert to `xarray`: `cf.Field.to_xarray`,
+  `cf.FieldList.to_xarray`, `cf.Domain.to_xarray`, and
+  `cf.DomainList.to_xarray`
+  (https://github.com/NCAS-CMS/cf-python/issues/933)
+* New output format for `cf.write` that creates an `xarray` dataset in
+  memory: ``'XARRAY'``
+  (https://github.com/NCAS-CMS/cf-python/issues/933)
 * New keyword parameter to `cf.Field.regrids` and `cf.Field.regridc`:
   ``mtol`` (https://github.com/NCAS-CMS/cf-python/issues/949)
 * Fix bug in `cf.read` that prevented some OPeNDAP URLS being read
@@ -10,6 +17,7 @@ Version 3.20.1
 * Fix bug that caused weighted `cf.collapse` and `cf.weights` to fail
   when a cell meausures has a size 1 axis
   (https://github.com/NCAS-CMS/cf-python/issues/952)
+* New optional dependency: ``xarray>=2026.2.0``
 
 ----